+++ pytypes.toml
@@ -1,0 +1,2 @@
+[pytypes]
+project = "PyTypes"
