+++ tracing/decorator_appender.py
@@ -52,6 +52,7 @@ class TracerDecoratorAppender:
             self.decorator_appended_file_paths.append(file_path_with_appended_decorators)
 
     def execute_decorator_appended_files(self):
+        """Executes the python files with the decorators appended to the pytest functions."""
         for decorator_appended_file_path in self.decorator_appended_file_paths:
             global_variables = {"__file__": decorator_appended_file_path}
             with decorator_appended_file_path.open("r") as file:
